crypto/curve_my_grade/requirements.txt
PyCryptodome
ecdsa